feat: added phases
