+++ build.gradle.kts
@@ -25,7 +25,7 @@ repositories {
 }
 
 dependencies {
-	implementation("io.github.architectplatform:architect-api:1.2.3")
+	implementation("io.github.architectplatform:architect-api:1.3.1")
 	ksp("io.micronaut:micronaut-http-validation")
 	ksp("io.micronaut.serde:micronaut-serde-processor")
 	implementation("io.micronaut.kotlin:micronaut-kotlin-runtime")
