+++ src/main/kotlin/io/github/architectplatform/engine/modules/tasks/publish/PublishCommand.kt
@@ -1,0 +1,21 @@
+package io.github.architectplatform.engine.modules.tasks.publish
+
+import io.github.architectplatform.api.command.Command
+import io.github.architectplatform.api.command.CommandRequest
+import io.github.architectplatform.api.tasks.publish.PublishTask
+import jakarta.inject.Singleton
+
+@Singleton
+class PublishCommand(val tasks: List<PublishTask>): Command<PublishCommandResult> {
+	override val name: String = "publish"
+
+	override fun execute(request: CommandRequest): PublishCommandResult {
+		println("Executing publish command")
+		val results = tasks.map { task ->
+			println("Running task: ${task.name}")
+			task.execute(request)
+		}
+		println("Run command completed")
+		return PublishCommandResult.Success(results)
+	}
+}
