+++ src/main/kotlin/io/github/architectplatform/engine/modules/tasks/release/ReleaseCommandResult.kt
@@ -1,0 +1,23 @@
+package io.github.architectplatform.engine.modules.tasks.release
+
+import io.github.architectplatform.api.command.CommandResult
+import io.github.architectplatform.api.tasks.release.ReleaseTaskResult
+
+interface ReleaseCommandResult : CommandResult {
+	/**
+	 * The result of the precommit command.
+	 */
+	val results: List<ReleaseTaskResult>
+
+	data class Success(
+		override val results: List<ReleaseTaskResult>,
+		override val success: Boolean = true,
+		override val output: String = "",
+	) : ReleaseCommandResult
+
+	data class Failure(
+		override val results: List<ReleaseTaskResult>,
+		override val success: Boolean = false,
+		override val output: String = "",
+	) : ReleaseCommandResult
+}
