--- src/main/kotlin/io/github/architectplatform/engine/tasks/init/InitCommandResult.kt
@@ -1,23 +1,0 @@
-package io.github.architectplatform.engine.tasks.init
-
-import io.github.architectplatform.api.command.CommandResult
-import io.github.architectplatform.api.tasks.init.InitTaskResult
-
-interface InitCommandResult : CommandResult {
-	/**
-	 * The result of the build command.
-	 */
-	val results: List<InitTaskResult>
-
-	data class Success(
-		override val results: List<InitTaskResult>,
-		override val success: Boolean = true,
-		override val output: String = "",
-	) : InitCommandResult
-
-	data class Failure(
-		override val results: List<InitTaskResult>,
-		override val success: Boolean = false,
-		override val output: String = "",
-	) : InitCommandResult
-}
