+++ src/main/resources/hooks/commit-msg
@@ -1,0 +1,4 @@
+echo "Running commit-msg hook..."
+architect commit-msg "$@"
+
+
