+++ src/main/resources/hooks/pre-commit
@@ -1,0 +1,4 @@
+echo "Running pre-commit hook..."
+architect pre-commit "$@"
+
+
