+++ src/main/resources/hooks/pre-push
@@ -1,0 +1,4 @@
+echo "Running pre-push hook..."
+architect pre-push "$@"
+
+
